Fix automatic repo detection if bad mirror
wget might not exit with error for a bad mirror (depends on the
server) in which case it might appear to be very fast (e.g. because
it is downloading a small redirected html file). This simple check
makes sure the file downloaded is the same as the second fastest repo
and if not it uses the second fastest repo.

This fix thus assumes that the second fastest repo is less likely
to be corrupt than the first repo. I have found this to be true
in practice.